Move all tests from `CollectionAssertionSpecs` to the appropriate files
